requirements.txt
google-cloud
botocore
aws-cdk.core
google-cloud-storage
google-auth
oauth2client
# Azure SDK 
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-network
azure-identity
azure-mgmt-resource 
tqdm 
